interfaces update

+++ src/main/java/fr/polytech/interfaces/fidelity/PointModifier.java
@@ -3,6 +3,7 @@ package fr.polytech.interfaces.fidelity;
 import fr.polytech.pojo.Customer;
 
 public interface PointModifier {
-    void incrementPoints(Customer customer, int price);
+    //points computed from the price
+    void incrementPoints(Customer customer, float price);
     void decrementPoints(Customer customer, int points);
 }
